feat: getAllLayerNames returns empty group + claire's review

+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/use_cases/regulatoryAreas/dtos/RegulatoryAreaGroupWithTotalDTO.kt
@@ -3,6 +3,6 @@ package fr.gouv.cacem.monitorenv.domain.use_cases.regulatoryAreas.dtos
 import fr.gouv.cacem.monitorenv.domain.entities.regulatoryArea.RegulatoryAreaEntity
 
 data class RegulatoryAreaGroupWithTotalDTO(
-    val total: Int,
+    val total: Long,
     val group: RegulatoryAreaEntity,
 )
